Refactor: Alert 영역 파일로 분리 및 접근제어 internal로 변경
